create an interface for GPGVerifier
