+++ key_verification/key_verify.go
@@ -4,13 +4,14 @@ import (
 	"context"
 
 	"github.com/ProtonMail/gopenpgp/v2/crypto"
+	"github.com/opentofu/libregistry/internal/gpg_verifier"
 	"github.com/opentofu/libregistry/types/provider"
 )
 
 func (kv keyVerification) VerifyKey(ctx context.Context, key *crypto.Key, namespace string) error {
 	namespace = provider.NormalizeNamespace(namespace)
 
-	signingKeyRing, err := parseKeyRing(key)
+	gpgVerifier, err := gpg_verifier.New(key)
 	if err != nil {
 		return err
 	}
@@ -37,7 +38,7 @@ func (kv keyVerification) VerifyKey(ctx context.Context, key *crypto.Key, namesp
 				return err
 			}
 
-			if err := validateDetachedSignature(signingKeyRing, shaSumContents, shaSumSigContents); err != nil {
+			if err := gpgVerifier.ValidateSignature(shaSumContents, shaSumSigContents); err != nil {
 				return err
 			}
 		}
